# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/seed-legacy":{"post":{"operationId":"AppController_seedLegacy","parameters":[],"responses":{"201":{"description":""}},"tags":["App"]}},"/events":{"post":{"operationId":"EventsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":""}},"tags":["Events"]},"get":{"operationId":"EventsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/{id}/full-structure":{"get":{"operationId":"EventsController_getFullStructure","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/{id}":{"get":{"operationId":"EventsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]},"patch":{"operationId":"EventsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]},"delete":{"operationId":"EventsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/{id}/session-speakers":{"get":{"operationId":"EventsController_getSessionSpeakers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/{id}/session-speakers/sort":{"patch":{"operationId":"EventsController_updateSessionSpeakersSort","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionSpeakersSortDto"}}}},"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/{id}/announcements":{"get":{"operationId":"EventsController_getAnnouncements","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]},"patch":{"operationId":"EventsController_updateAnnouncements","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAnnouncementsDto"}}}},"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/{id}/memo-template":{"get":{"operationId":"EventsController_getMemoTemplate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Events"]},"patch":{"operationId":"EventsController_updateMemoTemplate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMemoTemplateDto"}}}},"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/session-speakers/{id}":{"patch":{"operationId":"EventsController_updateSessionSpeaker","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionSpeakerDto"}}}},"responses":{"200":{"description":""}},"tags":["Events"]}},"/events/public/speaker-memo/{hash}":{"get":{"operationId":"EventsController_getPublicSpeakerMemo","parameters":[{"name":"hash","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Events"]}},"/public/events/{id}/website-data":{"get":{"operationId":"PublicEventsController_getWebsiteData","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["PublicEvents"]}},"/public/events/{id}/schedule":{"get":{"operationId":"PublicEventsController_getSchedule","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["PublicEvents"]}},"/public/events/{id}/speakers":{"get":{"operationId":"PublicEventsController_getSpeakers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["PublicEvents"]}},"/public/events/{id}/sponsors":{"get":{"operationId":"PublicEventsController_getSponsors","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["PublicEvents"]}},"/halls":{"post":{"operationId":"HallsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHallDto"}}}},"responses":{"201":{"description":""}},"summary":"Create a new Hall","tags":["halls"]},"get":{"operationId":"HallsController_findAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Get all Halls","tags":["halls"]}},"/halls/{id}":{"get":{"operationId":"HallsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get a specific Hall by ID","tags":["halls"]},"patch":{"operationId":"HallsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateHallDto"}}}},"responses":{"200":{"description":""}},"summary":"Update a specific Hall by ID","tags":["halls"]},"delete":{"operationId":"HallsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Delete a specific Hall by ID","tags":["halls"]}},"/tracks":{"post":{"operationId":"TracksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrackDto"}}}},"responses":{"201":{"description":""}},"tags":["Tracks"]},"get":{"operationId":"TracksController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Tracks"]}},"/tracks/{id}":{"get":{"operationId":"TracksController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tracks"]},"patch":{"operationId":"TracksController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrackDto"}}}},"responses":{"200":{"description":""}},"tags":["Tracks"]},"delete":{"operationId":"TracksController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Tracks"]}},"/sessions":{"post":{"operationId":"SessionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["Sessions"]},"get":{"operationId":"SessionsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/sessions/{id}":{"get":{"operationId":"SessionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Sessions"]},"patch":{"operationId":"SessionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionDto"}}}},"responses":{"200":{"description":""}},"tags":["Sessions"]},"delete":{"operationId":"SessionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/speakers":{"post":{"operationId":"SpeakersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSpeakerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerEntity"}}}}},"summary":"Create a new speaker","tags":["speakers"]},"get":{"operationId":"SpeakersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpeakerEntity"}}}}}},"summary":"Get all speakers","tags":["speakers"]}},"/speakers/import-legacy":{"post":{"operationId":"SpeakersController_importLegacy","parameters":[],"responses":{"201":{"description":""}},"summary":"Import legacy speakers","tags":["speakers"]}},"/speakers/sync-photos":{"post":{"operationId":"SpeakersController_syncPhotos","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":""}},"summary":"Sync speaker photos","tags":["speakers"]}},"/speakers/{id}":{"get":{"operationId":"SpeakersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerEntity"}}}}},"summary":"Get a single speaker by ID","tags":["speakers"]},"patch":{"operationId":"SpeakersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSpeakerDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerEntity"}}}}},"summary":"Update an existing speaker","tags":["speakers"]},"delete":{"operationId":"SpeakersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerEntity"}}}}},"summary":"Delete a speaker","tags":["speakers"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/{id}":{"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/change-password":{"post":{"operationId":"UsersController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/managers":{"get":{"operationId":"UsersController_findManagersDropdown","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/profile":{"put":{"operationId":"UsersController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/uploads/speaker-photo":{"post":{"operationId":"UploadsController_uploadSpeakerPhoto","parameters":[],"responses":{"201":{"description":""}},"tags":["Uploads"]}},"/uploads/logo":{"post":{"operationId":"UploadsController_uploadLogo","parameters":[],"responses":{"201":{"description":""}},"tags":["Uploads"]}},"/uploads/exact-logo/{filename}":{"post":{"operationId":"UploadsController_uploadExactLogo","parameters":[{"name":"filename","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Uploads"]}},"/uploads/presentation":{"post":{"operationId":"UploadsController_uploadPresentation","parameters":[],"responses":{"201":{"description":""}},"tags":["Uploads"]}},"/uploads/photos/{filename}":{"get":{"operationId":"UploadsController_servePhoto","parameters":[{"name":"filename","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Uploads"]}},"/uploads/logos/{filename}":{"get":{"operationId":"UploadsController_serveLogo","parameters":[{"name":"filename","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Uploads"]}},"/uploads/presentations/{filename}":{"get":{"operationId":"UploadsController_servePresentation","parameters":[{"name":"filename","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Uploads"]}},"/exports/schedule/{eventId}":{"get":{"operationId":"ExportsController_exportSchedule","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Exports"]}},"/exports/import/{eventId}":{"post":{"operationId":"ExportsController_importSchedule","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["Exports"]}},"/sponsors/event/{eventId}":{"get":{"operationId":"SponsorsController_findAllByEvent","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get sponsors by event ID","tags":["sponsors"]}},"/sponsors/all":{"get":{"operationId":"SponsorsController_findAll","parameters":[],"responses":{"200":{"description":""}},"summary":"Get all sponsors (Brands)","tags":["sponsors"]}},"/sponsors":{"post":{"operationId":"SponsorsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSponsorDto"}}}},"responses":{"201":{"description":""}},"summary":"Create a new sponsor (Brand)","tags":["sponsors"]}},"/sponsors/{id}":{"patch":{"operationId":"SponsorsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSponsorDto"}}}},"responses":{"200":{"description":""}},"summary":"Update an existing sponsor","tags":["sponsors"]},"delete":{"operationId":"SponsorsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Delete a sponsor","tags":["sponsors"]}},"/sponsors/{id}/attach/{eventId}":{"post":{"operationId":"SponsorsController_attachToEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"eventId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"summary":"Attach sponsor to an event","tags":["sponsors"]}},"/sponsors/{id}/detach/{eventId}":{"delete":{"operationId":"SponsorsController_detachFromEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"eventId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Detach sponsor from an event","tags":["sponsors"]}},"/sponsors/public/approval/{hash}":{"get":{"operationId":"SponsorsController_getPublicApprovalInfo","parameters":[{"name":"hash","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get public approval info for sponsor","tags":["sponsors"]},"patch":{"operationId":"SponsorsController_submitPublicApproval","parameters":[{"name":"hash","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicApprovalDto"}}}},"responses":{"200":{"description":""}},"summary":"Submit public approval for sponsor","tags":["sponsors"]}},"/sponsors/import-legacy":{"post":{"operationId":"SponsorsController_importLegacyBrands","parameters":[],"responses":{"201":{"description":""}},"summary":"Import legacy brands from previous CRM","tags":["sponsors"]}},"/sponsors/fix-brands":{"get":{"operationId":"SponsorsController_fixBrands","parameters":[],"responses":{"200":{"description":""}},"summary":"Run internal fix for brands","tags":["sponsors"]}},"/metrics":{"get":{"operationId":"PrometheusController_index","parameters":[],"responses":{"200":{"description":""}},"tags":["Prometheus"]}},"/market-segments/tree":{"get":{"operationId":"MarketSegmentsController_getTree","parameters":[],"responses":{"200":{"description":""}},"tags":["MarketSegments"]}},"/market-segments":{"get":{"operationId":"MarketSegmentsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["MarketSegments"]},"post":{"operationId":"MarketSegmentsController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["MarketSegments"]}},"/market-segments/{id}":{"patch":{"operationId":"MarketSegmentsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["MarketSegments"]},"delete":{"operationId":"MarketSegmentsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["MarketSegments"]}},"/market-segments/seed":{"post":{"operationId":"MarketSegmentsController_seed","parameters":[],"responses":{"201":{"description":""}},"tags":["MarketSegments"]}},"/v1/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}}},"info":{"title":"UPGRADE CRM API","description":"API documentation for UPGRADE CRM","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateEventDto":{"type":"object","properties":{}},"SpeakerSortUpdateItem":{"type":"object","properties":{"id":{"type":"number"},"sortOrder":{"type":"number"}},"required":["id","sortOrder"]},"UpdateSessionSpeakersSortDto":{"type":"object","properties":{"updates":{"type":"array","items":{"$ref":"#/components/schemas/SpeakerSortUpdateItem"}}},"required":["updates"]},"UpdateAnnouncementsDto":{"type":"object","properties":{"updates":{"type":"array","items":{"type":"object"}}},"required":["updates"]},"UpdateMemoTemplateDto":{"type":"object","properties":{"template":{"type":"string"}},"required":["template"]},"UpdateSessionSpeakerDto":{"type":"object","properties":{"role":{"type":"string"},"status":{"type":"string"},"sortOrder":{"type":"number"},"needs_zoom":{"type":"boolean"},"hasPresentation":{"type":"boolean"},"managerComment":{"type":"string"},"programThesis":{"type":"string"},"newsletterQuote":{"type":"string"},"presenceStatus":{"type":"string"},"companySnapshot":{"type":"string"},"positionSnapshot":{"type":"string"},"notifiedEmail":{"type":"boolean"},"notifiedTg":{"type":"boolean"},"exportToWebsite":{"type":"boolean"}}},"CreateHallDto":{"type":"object","properties":{"eventId":{"type":"number","description":"ID of the Event this hall belongs to","example":1},"name":{"type":"string","description":"Name of the Hall","example":"Main Conference Room"},"capacity":{"type":"number","description":"Maximum capacity of the hall","example":500},"sortOrder":{"type":"number","description":"Sorting order for UI display","example":10}},"required":["eventId","name"]},"UpdateHallDto":{"type":"object","properties":{}},"CreateTrackDto":{"type":"object","properties":{}},"UpdateTrackDto":{"type":"object","properties":{}},"CreateSessionDto":{"type":"object","properties":{}},"UpdateSessionDto":{"type":"object","properties":{}},"CreateSpeakerDto":{"type":"object","properties":{}},"SpeakerEntity":{"type":"object","properties":{"id":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"company":{"type":"object"},"position":{"type":"object"},"email":{"type":"object"},"phone":{"type":"object"},"telegram":{"type":"object"},"photoUrl":{"type":"object"},"bio":{"type":"object"},"isSponsor":{"type":"boolean"},"internalComment":{"type":"object"},"hasAssistant":{"type":"boolean"},"assistantName":{"type":"object"},"assistantContact":{"type":"object"}},"required":["id","firstName","lastName","isSponsor","hasAssistant"]},"UpdateSpeakerDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"UpdatePasswordDto":{"type":"object","properties":{}},"UpdateProfileDto":{"type":"object","properties":{}},"CreateSponsorDto":{"type":"object","properties":{"name":{"type":"string","description":"Название бренда/спонсора"},"shortDescription":{"type":"string","description":"Короткое описание"},"websiteUrl":{"type":"string","description":"Ссылка на веб-сайт"},"publicEmail":{"type":"string","description":"Публичный email"},"publicPhone":{"type":"string","description":"Публичный телефон"},"logoUrl":{"type":"string","description":"Ссылка на логотип"},"description":{"type":"string","description":"Подробное описание"},"catalogDescription":{"type":"string"},"serviceCardDescription":{"type":"string"},"city":{"type":"string"},"employeeCount":{"type":"number"},"annualTurnover":{"type":"string"},"telegram":{"type":"string"},"whatsapp":{"type":"string"},"contactName":{"type":"string"},"contactEmail":{"type":"string"},"cfoName":{"type":"string"},"cases":{"type":"object","description":"Кейсы (статьи или презентации)"},"status":{"type":"string","default":"pending"},"rejectionReason":{"type":"string"},"assignedManagerId":{"type":"number"},"exportToWebsite":{"type":"boolean"},"logoFileUrl":{"type":"string"}},"required":["name"]},"UpdateSponsorDto":{"type":"object","properties":{"name":{"type":"string","description":"Название бренда/спонсора"},"shortDescription":{"type":"string","description":"Короткое описание"},"websiteUrl":{"type":"string","description":"Ссылка на веб-сайт"},"publicEmail":{"type":"string","description":"Публичный email"},"publicPhone":{"type":"string","description":"Публичный телефон"},"logoUrl":{"type":"string","description":"Ссылка на логотип"},"description":{"type":"string","description":"Подробное описание"},"catalogDescription":{"type":"string"},"serviceCardDescription":{"type":"string"},"city":{"type":"string"},"employeeCount":{"type":"number"},"annualTurnover":{"type":"string"},"telegram":{"type":"string"},"whatsapp":{"type":"string"},"contactName":{"type":"string"},"contactEmail":{"type":"string"},"cfoName":{"type":"string"},"cases":{"type":"object","description":"Кейсы (статьи или презентации)"},"status":{"type":"string","default":"pending"},"rejectionReason":{"type":"string"},"assignedManagerId":{"type":"number"},"exportToWebsite":{"type":"boolean"},"logoFileUrl":{"type":"string"}}},"PublicApprovalDto":{"type":"object","properties":{"status":{"type":"string"},"rejectionReason":{"type":"string"}},"required":["status"]}}}}